refactor(FeedComponent): Alertpop 열고닫는 react훅 파일 분리
