+++ src/apps/front/src/features/feed/Explore.tsx
@@ -5,10 +5,10 @@ const Explore = () => {
   return (
     <div>
       <main>
-        <Lnb/>
+        <Lnb />
         <FeedLayout />
       </main>
-      <BigGnb/>
+      <BigGnb />
     </div>
   );
 };
