+++ src/apps/front/src/features/feed/_feed.scss
@@ -11,4 +11,4 @@ main {
 @import './components/Lnb.scss';
 @import './components/post.scss';
 @import './components/FeedLayout.scss';
-@import './components/Alertpop.scss'
+@import './components/Alertpop.scss';
