+++ src/apps/front/src/features/feed/components/Alertpop.scss
