+++ src/apps/front/src/features/feed/components/Alertpop.tsx
@@ -1,13 +1,18 @@
 
 
-const Alertpop = ({ isOpen, onClose }: { isOpen: boolean; onClose: () => void }) => {
+const Alertpop = ({isOpen,onClose,}: {isOpen: boolean;onClose: () => void;}) => {
+  
   // 배경 클릭 시 모달을 닫는 함수
-  const handleBgClick = (event: React.MouseEvent<HTMLDivElement, MouseEvent>) => {
+  const handleBgClick = (
+    event: React.MouseEvent<HTMLDivElement, MouseEvent>,
+  ) => {
     event.stopPropagation();
     onClose(); // 모달 닫기 함수 호출
   };
   // 모달 내용 클릭 시 이벤트 전파를 막는 함수
-  const handleContentClick = (event: React.MouseEvent<HTMLDivElement, MouseEvent>) => {
+  const handleContentClick = (
+    event: React.MouseEvent<HTMLDivElement, MouseEvent>,
+  ) => {
     event.stopPropagation(); // 이벤트 전파를 막음
   };
