+++ src/apps/front/src/features/feed/hooks/useModalAlert.tsx
@@ -1,0 +1,18 @@
+// useModalAlert.tsx
+import { useState } from 'react';
+
+const useModalAlert = () => {
+    const [isAlertOpen, setIsAlertOpen] = useState<boolean>(false);
+
+    const handleOpenAlert = () => {
+        setIsAlertOpen(true);
+    };
+
+    const handleCloseAlert = () => {
+        setIsAlertOpen(false);
+    };
+
+    return { isAlertOpen, handleOpenAlert, handleCloseAlert };
+};
+
+export default useModalAlert;
